+++ AAEmu.Game/Core/Packets/G2C/SCCompletedQuestsPacket.cs
@@ -8,13 +8,13 @@ public class SCCompletedQuestsPacket(CompletedQuest[] quests) : GamePacket(SCOff
 {
     public override PacketStream Write(PacketStream stream)
     {
-        stream.Write(quests.Length); // TODO max 200
+        stream.Write(quests.Length);
         foreach (var quest in quests)
         {
             var body = new byte[8];
             quest.Body.CopyTo(body, 0);
 
-            stream.Write(quest.Id); // idx
+            stream.Write((uint)quest.Id); // idx — client reads u32 (ushort Write desyncs / crash)
             stream.Write(body); // body
         }
         return stream;
